DB stuff now separate from Program.cs
Should now be a bit more efficient, will reuse builder across all connections
